fix: satisfy CI formatting and module-size gates
cargo fmt flagged a few wrapped assertions, and three legacy modules were at
their line ceilings, so the new attachment code was split into dedicated
files instead of growing them:

- RepositoryError moves to feedback/repository_error.rs.
- A new AttachmentPathResolver trait (a FeedbackRepository supertrait)
  carries the resolve methods; the storage impl lives in sqlite/resolve_ops.rs.
- The open_feedback_attachment command moves to a top-level open_attachment.rs,
  matching the existing clipboard_capture/screen_capture module layout.

Verified locally: cargo fmt, clippy -D warnings, cargo test --workspace
--all-targets, check:rust-size, check:terminology, svelte-check, vitest,
contracts:check and build:web all pass.
